docs: Overwrite unit tests when regenerate improves them
Improve the `regenerate_unit_tests` pipeline so that it compares the performance of the old vs the new unit tests, and if the new unit tests perform better (have a greater `num_passed`), then it overwrites the old tests with the new ones. This makes it easier to repeatedly run this pipeline until all of the unit tests pass.