feat: Limit processing to max `num-samples`
When generating the dataset, `{num_samples}` should be included in the template to limit the number of samples being generated.

But in case this is forgotten or doesn’t work, and for other pipelines we might want to run, use the `—num-samples` CLI parameter to limit the number of samples being processed.