feat: Add `message` as param to `log_item`
Can be used to log a message about an item, rather than properties of an item. If `message` is passed in, `properties` is ignored. If both are passed in, an error is raised.

Can be used in a pipeline to log an info message like:

```
log_item(message="Regenerating unit tests for {id}...")
```