feat: Add `DatasetPipeline`
Allows running a set of dataset actions. I debated whether this should just be in `Pipeline` itself, with `ItemPipeline` just being a different set of parameters. But I decided that, for now, the interfaces are just slightly different enough, it’s clearer if the specific type of pipeline is specified. But it wasn’t a strongly-opinionated decision.