feat: Add template support to all resolved params
Any parameter that uses `resolve_dataset_value` or `resolve_item_value` can now be passed a `Template`, which contains a string with variables to resolve.

For items, variables include the item’s `id`, each key in `data` and the `context`.

For datasets`, variables include each key in `metadata`, and the `context`.

This allows all parameters to have templates, and even functions which return templates.