fix: Use correct param types for default formatters
I used a JavaScript pattern, where the default value is assigned in the method signature, then when `undefined` is passed in, it just uses that value.

However, apparently Python doesn’t work like that. There appears no easy way to pass in a value for `formatters` which says to use the default value. So instead, I check if `formatters` is None and then assign the default formatters.

If no formatters are to be used, an empty dict can always be passed in.