feat: Remove implicit template formatting
All template formatting, aside from special formatting like prompts, will be done explicitly via the `Template` class.

This allows this formatting to be done for any parameter, without having to try to figure out which parameters have implicit formatting and which don’t.

Note that I tried auto-resolving formats during `resolve_item_value`. But that caused issues with parameters that were passing in code, which could have false parameters, or parameters not intended to be replaced (e.g., code which contains a template string `f”Error is: {error}”` which is intended to be resolved during runtime).

I concluded that the slight overhead of explicitly wrapping templates in `Template` is better.