feat: Use a structured context with child contexts
To make it easier to run child pipelines, it’ll be easier if `Context` is structured and not just wrapping a `dict`. This sets up `Context` with the key attributes it needs and the ability to create child contexts.

Because `Context` and `Pipeline` now have circular references, we need to jump through some hoops to make this work. I tried using `from __future__ import annotations`, but I couldn’t get it to work. I think because both classes reference the other class in their method signatures, so they are being referenced when the module is first loaded. Adding `type: ignore` solves the problem for now, but if we can find a way to clean this up later, that’d be great.