fix: `context` resolution in `if_item` expressions
Since we’re using `eval` for these, the attribute-like syntax used (e.g., `context.log_level`) was not being resolved.

Since the pipeline files are meant to be declarative, and shouldn’t care about whether something is a key or an attribute, we want these both to resolve the same.