feat(examples): Save chat logs to sample subfolders
Organize the chat logs by subfolder based on the sample ID. This makes it easier to see all the logs for a single sample. Also, add the timestamp from the current generation of the sample, so we can see when multiple attempts are made to generate a sample.

Note: At some point, this should also take into account each pass (i.e., use the `iteration` in the filename) so we don’t overwrite logs with multiple passes.