feat: Log a warning if we hit the max tokens
Previously, I was running into issues with Sonnet 3.7 where I was hitting the max tokens. I created rough code to handle continutations in this scenario, but I’m not seeing the behavior right now, and Sonnet may be expanding the output token max soon, so it’s not worth working on right now. For now, just warn when this occurs, and we can address it then.