fix: Output files relative to current working dir
While using the project root as the output root works when pipelines are being run from within this project, if this project is used as a package within another project, then the paths are wrong.

Since we were always running the `dataset-foundry` command from the project root anyway, this doesn’t change the effective paths when used within this project, only when used in other projects.