feat: Rebuild Docker image when Dockerfile changes
It’s hard to iterate on the Docker image if I keep having to hack around getting the image rebuilt. This makes it work like you’d expect: when the Dockerfile is changed, the image is rebuilt (or, more specifically, when the Dockerfile’s modification date is more recent than the image build date, then we rebuild).