feat: Collect stdout & stderr separately
Often it can be convenient to see if any errors were printed during a run by checking `stderr`. The old code combined both `stdout` and `stderr` into a single log stream, which made this impossible. This new code fixes this.

Because it can also be useful to see the interweaved logs, we’re still collecting a stream of both `stdout` and `stderr` together. This is what ges printed to the terminal if `stream_logs` is set.