refactor: Pass `prompt`, not `prompt_file` to agent
Later we may want to do multiple runs of the same repo with different prompts (e.g., to give the agent feedback on the repo they generated and have it update the code). Providing the prompt via `PROMPT.md` makes this awkward and error-prone.

Instead, pass it as an environment variable, which can then be used in the agent command.