chore: Remove `_metadata` key, insert `attempt`
Now that the `result` contains almost all of the same data as in `_metadata`, remove that key. To ensure the attempt count remains, insert it into the context given via the agent inputs. This is a bit hackish, but minor, and I don’t want to spend time redoing the entire interface just to add this one property..