fix: Stop sending firewall config updates to stderr
I had these lines in there to test whether stdout and stderr were properly interleaving in the logs captured by the container manager. However, for production, we want stderr to be empty unless there really were errors.

Note: We can’t convert these to stdout messages because stdout is what is being used to return the list of domains.