fix: Missing `contents` key resolves to `item.data`
If the `save_item` action is configured with a `Key` that points to a path that doesn’t exist underneath `item.data`, then the old code would return `item.data`, which is not what we want.

The new code fixes it so that any items that resolve to `None` get saved as an empty string.