feat: Add `items_key` parameter to `load_dataset`
Enables loading the data items from a specific key within a file, rather than treating the entire file as the list of items.

Specifically, this allows for the new format for `scenarios.yaml` that I’m working on that has a `metadata` key for provenance info and a `scenarios` key that contains a list of scenarios.