fix: Ensure timeout works with infinite loops
If the Docker container is running code that is in a tight infinite loop or is otherwise hung, then the asyncio timeout won’t fire.

By running the command using the `timeout` command, we ensure the timeout is triggered by the container OS, rather than us trying to detect the timeout outside of the container.

I left in the container timeout as a backup timeout, though we probably won’t need it.