fix: Allow `False`, `0` and empty collections for required
If a parameter for a dataset or item action is required, it should be valid to pass in a `False`, `0` or empty collection. The only things that should be disallowed are `None` and an empty string.

I was considering allowing an empty string as well, but there are so many places where we’re looking for a dict key, and I want to make sure that these are reject if they resolve to an empty string.