feat: Add `pipeline_log_handler` for per-item logs
When installed, redirects the logs for each item into the `logs` property for that item in the pipeline service. This allows each log to be displayed in separate views when the display supports it. If no item is active in the current context, then logs to the console service instead.