feat: Add `find_first` collection utility
I hate when code is hard to read, and while I’ve heard the `next(…)` syntax is idiomatic to get the first item in an array, it’s completely non-intuitive to me and hard to read. So here’s a simple `find_first` function that makes the intent easier to understand.