fix: Prevent context being too large
In `agent_runner`, we inject the context data into the environment. However, if the context data contains `item.data`, which then contains the output of test runs, this can exceed the maximum size of the environment variables and cause the sandbox not to initialize properly.

Since we’re injecting the item ID separately, there’s no need to include this in the context data either.