feat: Add `max-items` argument to CLI
Specifies the maximum number of concurrent items that can be processed.

This commit removes the pipeline options and moves the parameter into the pipeline `params`. This makes it easier to pass through from the command-line, and I was always a bit skeptical of adding an `options` field when we had params.

Note that the consequence of this is that, in theory, a pipeline can alter the concurrency of items (if it was processing items in a nested manner). In practice, this probably won’t ever happen (famous last words).