feat: Add `—limit` separate from `—num-samples`
After a dataset is generated, we usually want to process every sample in that dataset. `num-samples` was being overloaded to both specify the initial number of samples to generate and to limit the samples processed by other pipelines. This commit separates those two out, with `limit` being used to process only a specific number of samples and `num-samples` being reserved when generating new samples.