fix: Improve debugging when tests don’t run
If tests don’t run for any reason (e.g., a setup repo command is set to run before running the tests and it fails), then we don’t want to generically set the number of passing tests to 1. Rather, set it to 0 and trigger additional debugging information.

Also, if the log level is debug, include stdout in the string representation of `UnitTestResult`.